Added Faction Starting Army Counter
Squashed bug mentioned in previous commit. UnitFile is now contains the
list of units, instead of reading them individually. UnitFactNode is a
node for the starting army table. Currently doesn't have a home, but
will probably just be a reference object for the ViewModel as it only
act a consolidation of pointers to UnitDir in UnitFile.